refact(RHINENG-4688): refact models.py classes into separate files
