+++ app/models/schemas.py
@@ -1,0 +1,346 @@
+from copy import deepcopy
+
+from jsonschema import ValidationError as JsonSchemaValidationError
+from jsonschema import validate as jsonschema_validate
+from jsonschema.validators import Draft4Validator
+from marshmallow import EXCLUDE
+from marshmallow import Schema as MarshmallowSchema
+from marshmallow import ValidationError as MarshmallowValidationError
+from marshmallow import fields
+from marshmallow import post_load
+from marshmallow import pre_load
+from marshmallow import validate as marshmallow_validate
+from marshmallow import validates
+from marshmallow import validates_schema
+
+from app.models.constants import MAX_CANONICAL_FACTS_VERSION
+from app.models.constants import MIN_CANONICAL_FACTS_VERSION
+from app.models.constants import TAG_KEY_VALIDATION
+from app.models.constants import TAG_NAMESPACE_VALIDATION
+from app.models.constants import TAG_VALUE_VALIDATION
+from app.models.constants import ZERO_MAC_ADDRESS
+from app.models.constants import ProviderType
+from app.models.host import Host
+from app.models.host import LimitedHost
+from app.models.system_profile_normalizer import SystemProfileNormalizer
+from app.validators import check_empty_keys
+from app.validators import verify_ip_address_format
+from app.validators import verify_mac_address_format
+from app.validators import verify_satellite_id
+from app.validators import verify_uuid_format
+
+
+class DiskDeviceSchema(MarshmallowSchema):
+    device = fields.Str(validate=marshmallow_validate.Length(max=2048))
+    label = fields.Str(validate=marshmallow_validate.Length(max=1024))
+    options = fields.Dict(validate=check_empty_keys)
+    mount_point = fields.Str(validate=marshmallow_validate.Length(max=2048))
+    type = fields.Str(validate=marshmallow_validate.Length(max=256))
+
+
+class RhsmSchema(MarshmallowSchema):
+    version = fields.Str(validate=marshmallow_validate.Length(max=255))
+
+
+class OperatingSystemSchema(MarshmallowSchema):
+    major = fields.Int()
+    minor = fields.Int()
+    name = fields.Str(validate=marshmallow_validate.Length(max=4))
+
+
+class YumRepoSchema(MarshmallowSchema):
+    id = fields.Str(validate=marshmallow_validate.Length(max=256))
+    name = fields.Str(validate=marshmallow_validate.Length(max=1024))
+    gpgcheck = fields.Bool()
+    enabled = fields.Bool()
+    base_url = fields.Str(validate=marshmallow_validate.Length(max=2048))
+
+
+class DnfModuleSchema(MarshmallowSchema):
+    name = fields.Str(validate=marshmallow_validate.Length(max=128))
+    stream = fields.Str(validate=marshmallow_validate.Length(max=128))
+
+
+class InstalledProductSchema(MarshmallowSchema):
+    name = fields.Str(validate=marshmallow_validate.Length(max=512))
+    id = fields.Str(validate=marshmallow_validate.Length(max=64))
+    status = fields.Str(validate=marshmallow_validate.Length(max=256))
+
+
+class NetworkInterfaceSchema(MarshmallowSchema):
+    ipv4_addresses = fields.List(fields.Str())
+    ipv6_addresses = fields.List(fields.Str())
+    state = fields.Str(validate=marshmallow_validate.Length(max=25))
+    mtu = fields.Int()
+    mac_address = fields.Str(validate=marshmallow_validate.Length(max=59))
+    name = fields.Str(validate=marshmallow_validate.Length(min=1, max=50))
+    type = fields.Str(validate=marshmallow_validate.Length(max=18))
+
+
+class FactsSchema(MarshmallowSchema):
+    namespace = fields.Str()
+    facts = fields.Dict(validate=check_empty_keys)
+
+
+class TagsSchema(MarshmallowSchema):
+    class Meta:
+        unknown = EXCLUDE
+
+    namespace = fields.Str(required=False, allow_none=True, validate=TAG_NAMESPACE_VALIDATION)
+    key = fields.Str(required=True, allow_none=False, validate=TAG_KEY_VALIDATION)
+    value = fields.Str(required=False, allow_none=True, validate=TAG_VALUE_VALIDATION)
+
+
+class CanonicalFactsSchema(MarshmallowSchema):
+    class Meta:
+        unknown = EXCLUDE
+
+    canonical_facts_version = fields.Integer(
+        required=False,
+        load_default=MIN_CANONICAL_FACTS_VERSION,
+        validate=marshmallow_validate.Range(min=MIN_CANONICAL_FACTS_VERSION, max=MAX_CANONICAL_FACTS_VERSION),
+    )
+    is_virtual = fields.Boolean(required=False)
+
+    insights_id = fields.Str(validate=verify_uuid_format)
+    subscription_manager_id = fields.Str(validate=verify_uuid_format)
+    satellite_id = fields.Str(validate=verify_satellite_id)
+    fqdn = fields.Str(validate=marshmallow_validate.Length(min=1, max=255))
+    bios_uuid = fields.Str(validate=verify_uuid_format)
+    ip_addresses = fields.List(fields.Str(validate=verify_ip_address_format))
+    mac_addresses = fields.List(
+        fields.Str(validate=verify_mac_address_format), validate=marshmallow_validate.Length(min=1)
+    )
+    provider_id = fields.Str(validate=marshmallow_validate.Length(min=1, max=500))
+    provider_type = fields.Str(validate=marshmallow_validate.Length(min=1, max=50))
+
+    @validates_schema
+    def validate_schema(self, data, **kwargs):
+        schema_version = data.get("canonical_facts_version")
+
+        if "mac_addresses" in data:
+            mac_addresses = data["mac_addresses"]
+            while ZERO_MAC_ADDRESS in mac_addresses:
+                from app.models import logger
+
+                logger.warning(f"Zero MAC address reported by: {data.get('reporter', 'Not Available')}")
+                mac_addresses.remove(ZERO_MAC_ADDRESS)
+            if not mac_addresses:
+                del data["mac_addresses"]
+
+        if schema_version > MIN_CANONICAL_FACTS_VERSION:
+            if "is_virtual" not in data:
+                raise MarshmallowValidationError(
+                    f"is_virtual is required for canonical_facts_version > {MIN_CANONICAL_FACTS_VERSION}."
+                )
+            if "mac_addresses" not in data:
+                raise MarshmallowValidationError(
+                    f"mac_addresses is required for canonical_facts_version > {MIN_CANONICAL_FACTS_VERSION}."
+                )
+            if data["is_virtual"]:
+                if "provider_id" not in data:
+                    raise MarshmallowValidationError(
+                        "provider_id and provider_type are required when is_virtual = True."
+                    )
+            else:
+                if "provider_id" in data:
+                    raise MarshmallowValidationError("provider_id is not allowed when is_virtual = False.")
+
+        provider_type = data.get("provider_type")
+        provider_id = data.get("provider_id")
+
+        if (provider_type and not provider_id) or (provider_id and not provider_type):
+            raise MarshmallowValidationError("provider_type and provider_id are both required.")
+
+        if provider_type and provider_type.lower() not in ProviderType.__members__.values():
+            raise MarshmallowValidationError(
+                f'Unknown Provider Type: "{provider_type}".  '
+                f"Valid provider types are: {', '.join([p.value for p in ProviderType])}."
+            )
+
+        if provider_id and provider_id.isspace():
+            raise MarshmallowValidationError("Provider id can not be just blank, whitespaces or tabs")
+
+
+class LimitedHostSchema(CanonicalFactsSchema):
+    class Meta:
+        unknown = EXCLUDE
+
+    display_name = fields.Str(validate=marshmallow_validate.Length(min=1, max=200))
+    ansible_host = fields.Str(validate=marshmallow_validate.Length(min=0, max=255))
+    account = fields.Str(validate=marshmallow_validate.Length(min=0, max=10))
+    org_id = fields.Str(required=True, validate=marshmallow_validate.Length(min=1, max=36))
+    facts = fields.List(fields.Nested(FactsSchema))
+    system_profile = fields.Dict()
+    tags = fields.Raw()
+    tags_alt = fields.Raw()
+    groups = fields.List(fields.Dict())
+
+    def __init__(self, system_profile_schema=None, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        cls = type(self)
+        if not hasattr(cls, "system_profile_normalizer"):
+            cls.system_profile_normalizer = SystemProfileNormalizer()
+        if system_profile_schema:
+            self.system_profile_normalizer = SystemProfileNormalizer(system_profile_schema=system_profile_schema)
+
+    @validates("tags")
+    def validate_tags(self, tags, data_key):  # noqa: ARG002, required for marshmallow validator functions
+        if isinstance(tags, list):
+            return self._validate_tags_list(tags)
+        elif isinstance(tags, dict):
+            return self._validate_tags_dict(tags)
+        else:
+            raise MarshmallowValidationError("Tags must be either an object or an array, and cannot be null.")
+
+    @staticmethod
+    def _validate_tags_list(tags):
+        TagsSchema(many=True).load(tags)
+        return True
+
+    @staticmethod
+    def _validate_tags_dict(tags):
+        for namespace, ns_tags in tags.items():
+            TAG_NAMESPACE_VALIDATION(namespace)
+            if ns_tags is None:
+                continue
+            if not isinstance(ns_tags, dict):
+                raise MarshmallowValidationError("Tags in a namespace must be an object or null.")
+
+            for key, values in ns_tags.items():
+                TAG_KEY_VALIDATION(key)
+                if values is None:
+                    continue
+                if not isinstance(values, list):
+                    raise MarshmallowValidationError("Tag values must be an array or null.")
+
+                for value in values:
+                    if value is None:
+                        continue
+                    if not isinstance(value, str):
+                        raise MarshmallowValidationError("Tag value must be a string or null.")
+                    TAG_VALUE_VALIDATION(value)
+
+        return True
+
+    @staticmethod
+    def _normalize_system_profile(normalize, data):
+        if "system_profile" not in data:
+            return data
+
+        system_profile = deepcopy(data["system_profile"])
+        normalize(system_profile)
+        return {**data, "system_profile": system_profile}
+
+    @staticmethod
+    def build_model(data, canonical_facts, facts, tags, tags_alt=None):
+        return LimitedHost(
+            canonical_facts=canonical_facts,
+            display_name=data.get("display_name"),
+            ansible_host=data.get("ansible_host"),
+            account=data.get("account"),
+            org_id=data.get("org_id"),
+            facts=facts,
+            tags=tags,
+            tags_alt=tags_alt if tags_alt else [],
+            system_profile_facts=data.get("system_profile", {}),
+            groups=data.get("groups", []),
+        )
+
+    @pre_load
+    def coerce_system_profile_types(self, data, **kwargs):
+        return self._normalize_system_profile(self.system_profile_normalizer.coerce_types, data)
+
+    @post_load
+    def filter_system_profile_keys(self, data, **kwargs):
+        return self._normalize_system_profile(self.system_profile_normalizer.filter_keys, data)
+
+    @validates("system_profile")
+    def system_profile_is_valid(self, system_profile, data_key):  # noqa: ARG002, required for marshmallow validator functions
+        try:
+            jsonschema_validate(
+                system_profile, self.system_profile_normalizer.schema, format_checker=Draft4Validator.FORMAT_CHECKER
+            )
+        except JsonSchemaValidationError as error:
+            raise MarshmallowValidationError(f"System profile does not conform to schema.\n{error}") from error
+
+        for dd_i, disk_device in enumerate(system_profile.get("disk_devices", [])):
+            if not check_empty_keys(disk_device.get("options")):
+                raise MarshmallowValidationError(f"Empty key in /system_profile/disk_devices/{dd_i}/options.")
+
+
+class HostSchema(LimitedHostSchema):
+    class Meta:
+        unknown = EXCLUDE
+
+    stale_timestamp = fields.AwareDateTime(required=True)
+    reporter = fields.Str(required=True, validate=marshmallow_validate.Length(min=1, max=255))
+
+    @staticmethod
+    def build_model(data, canonical_facts, facts, tags, tags_alt=None):
+        if tags_alt is None:
+            tags_alt = []
+        return Host(
+            canonical_facts,
+            data.get("display_name"),
+            data.get("ansible_host"),
+            data.get("account"),
+            data.get("org_id"),
+            facts,
+            tags,
+            tags_alt,
+            data.get("system_profile", {}),
+            data["stale_timestamp"],
+            data["reporter"],
+            data.get("groups", []),
+        )
+
+
+class PatchHostSchema(MarshmallowSchema):
+    ansible_host = fields.Str(validate=marshmallow_validate.Length(min=0, max=255))
+    display_name = fields.Str(validate=marshmallow_validate.Length(min=1, max=200))
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+
+class InputGroupSchema(MarshmallowSchema):
+    name = fields.Str(validate=marshmallow_validate.Length(min=1, max=255))
+    host_ids = fields.List(fields.Str(validate=verify_uuid_format))
+
+    @pre_load
+    def strip_whitespace_from_name(self, in_data, **kwargs):
+        if "name" in in_data:
+            in_data["name"] = in_data["name"].strip()
+
+        return in_data
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+
+class StalenessSchema(MarshmallowSchema):
+    conventional_time_to_stale = fields.Integer(validate=marshmallow_validate.Range(min=1, max=604800))
+    conventional_time_to_stale_warning = fields.Integer(validate=marshmallow_validate.Range(min=1, max=15552000))
+    conventional_time_to_delete = fields.Integer(validate=marshmallow_validate.Range(min=1, max=63072000))
+    immutable_time_to_stale = fields.Integer(validate=marshmallow_validate.Range(min=1, max=604800))
+    immutable_time_to_stale_warning = fields.Integer(validate=marshmallow_validate.Range(min=1, max=15552000))
+    immutable_time_to_delete = fields.Integer(validate=marshmallow_validate.Range(min=1, max=63072000))
+
+    @validates_schema
+    def validate_staleness(self, data, **kwargs):
+        staleness_fields = ["time_to_stale", "time_to_stale_warning", "time_to_delete"]
+        for host_type in (
+            "conventional",
+            "immutable",
+        ):
+            for i in range(len(staleness_fields) - 1):
+                for j in range(i + 1, len(staleness_fields)):
+                    if (
+                        data[(field_1 := f"{host_type}_{staleness_fields[i]}")]
+                        >= data[(field_2 := f"{host_type}_{staleness_fields[j]}")]
+                    ):
+                        raise MarshmallowValidationError(f"{field_1} must be lower than {field_2}")
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
