+++ app/models/system_profile_normalizer.py
@@ -1,0 +1,74 @@
+from collections import namedtuple
+from enum import Enum
+from os.path import join
+
+from connexion.utils import coerce_type
+from jsonschema import RefResolver
+from yaml import safe_load
+
+from app.models.constants import SPECIFICATION_DIR
+from app.models.constants import SYSTEM_PROFILE_SPECIFICATION_FILE
+
+
+class SystemProfileNormalizer:
+    class Schema(namedtuple("Schema", ("type", "properties", "items"))):
+        Types = Enum("Types", ("array", "object"))
+
+        @classmethod
+        def from_dict(cls, schema, resolver):
+            if "$ref" in schema:
+                _, schema = resolver.resolve(schema["$ref"])
+
+            filtered = {key: schema.get(key) for key in cls._fields}
+            return cls(**filtered)
+
+        @property
+        def schema_type(self):
+            return self.Types.__members__.get(self.type)
+
+    SOME_ARBITRARY_STRING = "property"
+
+    def __init__(self, system_profile_schema=None):
+        if system_profile_schema:
+            system_profile_spec = system_profile_schema
+        else:
+            specification = join(SPECIFICATION_DIR, SYSTEM_PROFILE_SPECIFICATION_FILE)
+            with open(specification) as file:
+                system_profile_spec = safe_load(file)
+
+        self.schema = {**system_profile_spec, "$ref": "#/$defs/SystemProfile"}
+        self._resolver = RefResolver.from_schema(system_profile_spec)
+
+    def filter_keys(self, payload, schema_dict=None):
+        if schema_dict is None:
+            schema_dict = self._system_profile_definition()
+
+        schema_obj = self.Schema.from_dict(schema_dict, self._resolver)
+        if schema_obj.schema_type == self.Schema.Types.object:
+            self._object_filter(schema_obj, payload)
+        elif schema_obj.schema_type == self.Schema.Types.array:
+            self._array_filter(schema_obj, payload)
+
+    def coerce_types(self, payload, schema_dict=None):
+        if schema_dict is None:
+            schema_dict = self._system_profile_definition()
+        coerce_type(schema_dict, payload, self.SOME_ARBITRARY_STRING)
+
+    def _system_profile_definition(self):
+        return self.schema["$defs"]["SystemProfile"]
+
+    def _object_filter(self, schema, payload):
+        if not schema.properties or type(payload) is not dict:
+            return
+
+        for key in payload.keys() - schema.properties.keys():
+            del payload[key]
+        for key in payload:
+            self.filter_keys(payload[key], schema.properties[key])
+
+    def _array_filter(self, schema, payload):
+        if not schema.items or type(payload) is not list:
+            return
+
+        for value in payload:
+            self.filter_keys(value, schema.items)
